Working branch

+++ src/main/java/hark/ecom/entities/Cart.java
@@ -2,17 +2,14 @@ package hark.ecom.entities;
 
 import hark.ecom.entities.enums.CartStatus;
 import hark.ecom.entities.products.CartItem;
-import hark.ecom.entities.products.Product;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
-import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
-import java.util.Set;
 
 @Entity
 @Getter
@@ -23,7 +20,7 @@ public class Cart {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "cart_id")
-    private int id;
+    private Long id;
 
     @Column(name = "cart_tracking_number")
     private String orderTrackingNumber;
